splitting ocp and mem handlers
