+++ koku/api/report/ocp/ocp_query_handler.py
@@ -14,7 +14,7 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
-"""AWS Query Handling for Reports."""
+"""OCP Query Handling for Reports."""
 import copy
 import datetime
 from collections import OrderedDict
@@ -36,11 +36,10 @@ from api.report.query_filter import QueryFilterCollection
 class OCPReportQueryHandler(ReportQueryHandler):
     """Handles report queries and responses for AWS."""
 
-    default_ordering = {'total': 'desc'}
     group_by_options = ['cluster', 'project', 'node']
 
     def __init__(self, query_parameters, url_data,
-                 tenant, **kwargs):
+                 tenant, default_ordering, **kwargs):
         """Establish AWS report query handler.
 
         Args:
@@ -50,9 +49,8 @@ class OCPReportQueryHandler(ReportQueryHandler):
             kwargs    (Dict): A dictionary for internal query alteration based on path
         """
         kwargs['provider'] = 'OCP'
-
         super().__init__(query_parameters, url_data,
-                         tenant, self.default_ordering,
+                         tenant, default_ordering,
                          self.group_by_options, **kwargs)
 
     def _ranked_list(self, data_list):
@@ -190,26 +188,6 @@ class OCPReportQueryHandler(ReportQueryHandler):
                                 reverse=reverse)
         return query_data
 
-    def _build_query(self, query_data, query_group_by):
-
-        if self._report_type == 'cpu':
-            cpu_usage = self._mapper._report_type_map.get('cpu_usage')
-            cpu_request = self._mapper._report_type_map.get('cpu_request')
-            cpu_limit = self._mapper._report_type_map.get('cpu_limit')
-            query_data = query_data.values(*query_group_by)\
-                .annotate(cpu_usage_core_hours=Sum(cpu_usage))\
-                .annotate(cpu_requests_core_hours=Sum(cpu_request))\
-                .annotate(cpu_limit=Sum(cpu_limit))
-
-        elif self._report_type == 'mem':
-            mem_usage = self._mapper._report_type_map.get('mem_usage')
-            mem_request = self._mapper._report_type_map.get('mem_request')
-            query_data = query_data.values(*query_group_by)\
-                .annotate(memory_usage_gigabytes=Sum(mem_usage))\
-                .annotate(memory_requests_gigabytes=Sum(mem_request))
-        
-        return query_data
-
     def execute_sum_query(self):
         """Execute query and return provided data when self.is_sum == True.
 
@@ -217,59 +195,7 @@ class OCPReportQueryHandler(ReportQueryHandler):
             (Dict): Dictionary response of query params, data, and total
 
         """
-
-        query_sum = {'value': 0}
-        data = []
-
-        q_table = self._mapper._operation_map.get('tables').get('query')
-        with tenant_context(self.tenant):
-            query = q_table.objects.filter(self.query_filter)
-            query_annotations = self._get_annotations()
-            query_data = query.annotate(**query_annotations)
-            group_by_value = self._get_group_by()
-            query_group_by = ['date'] + group_by_value
-
-            query_order_by = ('-date', )
-            # if self.order_field != 'delta':
-                # query_order_by += (self.order,)
-            query_data = self._build_query(query_data, query_group_by)
-            if self._mapper.count:
-                # This is a sum because the summary table already
-                # has already performed counts
-                query_data = query_data.annotate(count=Sum(self._mapper.count))
-
-            if self._limit and group_by_value:
-                rank_order = getattr(F(group_by_value.pop()), self.order_direction)()
-                dense_rank_by_total = Window(
-                    expression=DenseRank(),
-                    partition_by=F('date'),
-                    order_by=rank_order
-                )
-                query_data = query_data.annotate(rank=dense_rank_by_total)
-                query_order_by = query_order_by + ('rank',)
-
-            if self.order_field != 'delta':
-                query_data = query_data.order_by(*query_order_by)
-
-            if query.exists():
-                query_sum = self.calculate_total()
-
-            if self._delta:
-                query_data = self.add_deltas(query_data, query_sum)
-
-            is_csv_output = self._accept_type and 'text/csv' in self._accept_type
-
-            if is_csv_output:
-                if self._limit:
-                    data = self._ranked_list(list(query_data))
-                else:
-                    data = list(query_data)
-            else:
-                data = self._apply_group_by(list(query_data))
-                data = self._transform_data(query_group_by, 0, data)
-        self.query_sum = query_sum
-        self.query_data = data
-        return self._format_query_response()
+        pass
 
     def execute_query(self):
         """Execute query and return provided data.
@@ -290,36 +216,4 @@ class OCPReportQueryHandler(ReportQueryHandler):
             (dict) The aggregated totals for the query
 
         """
-        filt_collection = QueryFilterCollection()
-        total_filter = self._get_search_filter(filt_collection)
-
-        time_scope_value = self.get_query_param_data('filter',
-                                                     'time_scope_value',
-                                                     -10)
-        time_and_report_filter = Q(time_scope_value=time_scope_value)
-
-        if total_filter is None:
-            total_filter = time_and_report_filter
-        else:
-            total_filter = total_filter & time_and_report_filter
-
-        q_table = self._mapper._operation_map.get('tables').get('total')
-        total_query = q_table.objects.filter(total_filter)
-
-        total_dict = {}
-        if self._report_type == 'cpu':
-            cpu_usage_key = self._mapper._report_type_map.get('cpu_usage')
-            cpu_request_key = self._mapper._report_type_map.get('cpu_request')
-            cpu_usage_sum = total_query.aggregate(cpu_usage=Sum(cpu_usage_key))
-            cpu_request_sum = total_query.aggregate(cpu_request=Sum(cpu_request_key))
-            total_dict['cpu_usage_core_hours'] = cpu_usage_sum.get('cpu_usage')
-            total_dict['cpu_requests_core_hours'] = cpu_request_sum.get('cpu_request')
-        elif self._report_type == 'mem':
-            mem_usage_key = self._mapper._report_type_map.get('mem_usage')
-            mem_request_key = self._mapper._report_type_map.get('mem_request')
-            mem_usage_sum = total_query.aggregate(mem_usage=Sum(mem_usage_key))
-            mem_request_sum = total_query.aggregate(mem_request=Sum(mem_request_key))
-            total_dict['memory_usage_gigabytes'] = mem_usage_sum.get('mem_usage')
-            total_dict['memory_requests_gigabytes'] = mem_request_sum.get('mem_request')
-        
-        return total_dict
+        pass
